webui: extract ModelOption type into dedicated models.d.ts
